Set up Slack command
This command will eventually be the crux of a Slack app that parses a
list of songs from a Google spreadsheet doc and create a spotify
playlist.

For now, it just returns a canned response. Added gems so we can use
`binding.pry`.

Also added a dummy verify signature before action that will be used to
ensure that the request is indeed is from Slack.